+++ src/frontend/src/components/buttons/YesNoButton.tsx
@@ -1,7 +1,7 @@
 import { t } from '@lingui/core/macro';
 import { Badge, Skeleton } from '@mantine/core';
 
-import { isTrue } from '../../functions/conversion';
+import { isTrue } from '@lib/functions/Conversion';
 
 export function PassFailButton({
   value,
