+++ src/frontend/src/hooks/UseDashboardItems.tsx
@@ -3,6 +3,7 @@ import { useMemo } from 'react';
 
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { apiUrl } from '@lib/functions/Api';
+import { identifierString } from '@lib/functions/Conversion';
 import { api } from '../App';
 import type { DashboardWidgetProps } from '../components/dashboard/DashboardWidget';
 import DashboardWidgetLibrary from '../components/dashboard/DashboardWidgetLibrary';
@@ -12,7 +13,6 @@ import {
   PluginUIFeatureType
 } from '../components/plugins/PluginUIFeature';
 import RemoteComponent from '../components/plugins/RemoteComponent';
-import { identifierString } from '../functions/conversion';
 import { useGlobalSettingsState } from '../states/SettingsStates';
 import { useUserState } from '../states/UserState';
