+++ src/frontend/src/hooks/UseFilter.tsx
@@ -5,9 +5,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { useCallback, useMemo } from 'react';
 
+import { resolveItem } from '@lib/functions/Conversion';
 import type { TableFilterChoice } from '@lib/types/Filters';
 import { useApi } from '../contexts/ApiContext';
-import { resolveItem } from '../functions/conversion';
 
 type UseFilterProps = {
   url: string;
